Fix slider bug

+++ document-viewer/src/main/java/com/artifex/mupdf/viewer/LLDocumentActivity.kt
@@ -251,7 +251,7 @@ class LLDocumentActivity : ComponentActivity() {
                                                 pageCount = pageCount,
                                                 onPageChange = { newPageIndex ->
                                                     coroutineScope.launch {
-                                                        pagerState.scrollToPage(newPageIndex.toInt())
+                                                        pagerState.scrollToPage(newPageIndex)
                                                     }
                                                 }
                                             )
